+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate:
 
+### v4.39.0.0 (uncut)
+- **http4k-contract** : Support for HTTP callbacks in OpenApi3
+
 ### v4.38.0.1
-- **http4k-core** : Fix to header parsing to trim both start and end of 
+- **http4k-core** : [Fix] Header parsing to split correctly.
 
 ### v4.38.0.0
 - **http4k-*** : Upgrade some dependency versions, including Kotlin to 1.8.10.